+++ packages/dartway_core_server/lib/dartway_core_server.dart
@@ -48,6 +48,7 @@ export 'src/jobs/dw_job_queue.dart'
 export 'src/routes/dw_http_route.dart';
 export 'src/server/dw_app_server.dart';
 export 'src/server/dw_local_environment.dart' show DwLocalEnvironment;
+export 'src/server/dw_server_feature.dart';
 export 'src/server/dw_server_module.dart';
 export 'src/server/dw_server_settings.dart';
 export 'src/server/dw_startup_step.dart'