+++ template/dartway_starter_server/bin/seed_dev.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:dartway_core_server/dartway_core_server.dart';
 import 'package:dartway_starter_server/dartway_starter_server.dart';
-import 'package:dartway_starter_server/src/entities/people.dart';
+import 'package:dartway_starter_server/src/profile/profile_rows.dart';
 import 'package:dartway_starter_shared/dartway_starter_shared.dart';
 
 /// The code every seeded persona signs in with instead of a delivered one.